Add GA

+++ aws-infra/aws-up.sh
@@ -26,9 +26,14 @@ terraform apply -parallelism=1 -auto-approve
 # Update .kube/config context
 echo "Fetching AKS creds for kubectl"
 aws eks update-kubeconfig --region $TF_VAR_region --name $TF_VAR_name-eks
-# Rename contexts
-kubectl config delete-context $TF_VAR_name
-kubectl config rename-context arn:aws:eks:${TF_VAR_region}:${TF_VAR_aws_acccount}:cluster/${TF_VAR_name}-eks $TF_VAR_name
+
+# Grab AWS env variables
+if [[  -z "${IS_GITHUB_ACTIONS}" ]]; then
+  # Rename contexts
+  kubectl config delete-context $TF_VAR_name
+  kubectl config rename-context arn:aws:eks:${TF_VAR_region}:${TF_VAR_aws_acccount}:cluster/${TF_VAR_name}-eks $TF_VAR_name
+fi
+
 
 # Install ArgoCD
 echo "Installing ArgoCD"
